Implemented willbet.will_bet_user().
