+++ blackjack/model.py
@@ -14,45 +14,6 @@ from unicodedata import normalize
 from typing import Any, Iterable, Sequence
 
 
-# Base classes.
-class BaseEngine(ABC):
-    bet_max: int
-    bet_min: int
-
-    """The base class for the game engine."""
-    @abstractmethod
-    def deal(self) -> None:
-        """Dealing phase."""
-
-    @abstractmethod
-    def end(self) -> None:
-        """End of the hand."""
-
-    @abstractmethod
-    def new_game(self) -> None:
-        """Start a new game."""
-
-    @abstractmethod
-    def play(self) -> None:
-        """Play the hand."""
-
-    @abstractmethod
-    def restore(self, fname: str) -> None:
-        """Restore a game from a save."""
-
-    @abstractmethod
-    def save(self, fname: str) -> None:
-        """Save a game."""
-
-    @abstractmethod
-    def serialize(self) -> None:
-        """Serialize the game."""
-
-    @abstractmethod
-    def start(self) -> None:
-        """Start of the hand."""
-
-
 # Base validation classes.
 class _BaseDescriptor:
     """A basic data descriptor."""
@@ -330,3 +291,164 @@ class IsYes:
         if not isinstance(other, cls):
             return NotImplemented
         return self.value == other.value
+
+
+# Base classes.
+class EngineUI(ABC):
+    # General operation methods.
+    @abstractmethod
+    def end(self):
+        """End the UI."""
+
+    @abstractmethod
+    def reset(self):
+        """Restart the UI."""
+
+    @abstractmethod
+    def start(self):
+        """Start the UI."""
+
+    # Input methods.
+    @abstractmethod
+    def bet_prompt(self, bet_min: int, bet_max: int) -> Bet:
+        """Ask user for a bet.."""
+
+    @abstractmethod
+    def doubledown_prompt(self) -> IsYes:
+        """Ask user if they want to double down."""
+
+    @abstractmethod
+    def hit_prompt(self) -> IsYes:
+        """Ask user if they want to hit."""
+
+    @abstractmethod
+    def insure_prompt(self) -> IsYes:
+        """Ask user if they want to insure."""
+
+    @abstractmethod
+    def nextgame_prompt(self) -> IsYes:
+        """Ask user if they want to play another round."""
+
+    @abstractmethod
+    def split_prompt(self) -> IsYes:
+        """Ask user if they want to split."""
+
+    # Update methods.
+    @abstractmethod
+    def bet(self, player, bet):
+        """Player places initial bet."""
+
+    @abstractmethod
+    def cleanup(self):
+        """Clean up after the round ends."""
+
+    @abstractmethod
+    def deal(self, player, hand):
+        """Player receives initial hand."""
+
+    @abstractmethod
+    def doubledown(self, player, bet):
+        """Player doubles down."""
+
+    @abstractmethod
+    def flip(self, player, hand):
+        """Player flips a card."""
+
+    @abstractmethod
+    def hit(self, player, hand):
+        """Player hits."""
+
+    @abstractmethod
+    def insures(self, player, bet):
+        """Player insures their hand."""
+
+    @abstractmethod
+    def insurepay(self, player, bet):
+        """Insurance is paid to player."""
+
+    @abstractmethod
+    def joins(self, player):
+        """Player joins the game."""
+
+    @abstractmethod
+    def leaves(self, player):
+        """Player leaves the game."""
+
+    @abstractmethod
+    def loses(self, player):
+        """Player loses."""
+
+    @abstractmethod
+    def loses_split(self, player):
+        """Player loses on their split hand."""
+
+    @abstractmethod
+    def shuffles(self, player):
+        """The deck is shuffled."""
+
+    @abstractmethod
+    def splits(self, player, bet):
+        """Player splits their hand."""
+
+    @abstractmethod
+    def stand(self, player, hand):
+        """Player stands."""
+
+    @abstractmethod
+    def tie(self, player, bet):
+        """Player ties."""
+
+    @abstractmethod
+    def ties_split(self, player, bet):
+        """Player ties on their split hand."""
+
+    @abstractmethod
+    def update_count(self, count):
+        """Update the running card count in the UI."""
+
+    @abstractmethod
+    def wins(self, player, bet):
+        """Player wins."""
+
+    @abstractmethod
+    def wins_split(self, player, bet):
+        """Player wins on their split hand."""
+
+
+class BaseEngine(ABC):
+    bet_max: int
+    bet_min: int
+    ui: EngineUI
+
+    """The base class for the game engine."""
+    @abstractmethod
+    def deal(self) -> None:
+        """Dealing phase."""
+
+    @abstractmethod
+    def end(self) -> None:
+        """End of the hand."""
+
+    @abstractmethod
+    def new_game(self) -> None:
+        """Start a new game."""
+
+    @abstractmethod
+    def play(self) -> None:
+        """Play the hand."""
+
+    @abstractmethod
+    def restore(self, fname: str) -> None:
+        """Restore a game from a save."""
+
+    @abstractmethod
+    def save(self, fname: str) -> None:
+        """Save a game."""
+
+    @abstractmethod
+    def serialize(self) -> None:
+        """Serialize the game."""
+
+    @abstractmethod
+    def start(self) -> None:
+        """Start of the hand."""
